requirements.txt
Flask==3.0.0
requests==2.32.3
supabase==2.3.5
python-dotenv==1.0.0
